# openapi/cognigate-api.yaml
openapi: 3.1.0
info:
  title: Cognigate API
  description: |
    # Cognigate AI Governance API

    Cognigate provides AI governance, trust scoring, and policy enforcement for autonomous agents.

    ## Base URLs
    - **Production**: `https://cognigate.dev/v1`
    - **Staging**: `https://staging.cognigate.dev/v1`

    ## Authentication
    All endpoints require Bearer token authentication:
    ```
    Authorization: Bearer <your-api-key>
    ```

    ## Trust Tiers (BASIS Specification)
    | Tier | Score Range | Name | Description |
    |------|-------------|------|-------------|
    | T0 | 0-199 | Sandbox | Isolated, no external access |
    | T1 | 200-349 | Observed | Read-only, monitored |
    | T2 | 350-500 | Provisional | Basic operations, supervised |
    | T3 | 501-649 | Monitored | Standard operations with monitoring |
    | T4 | 650-799 | Standard | External API access, policy-governed |
    | T5 | 800-875 | Trusted | Cross-agent communication |
    | T6 | 876-950 | Certified | Admin tasks, minimal oversight |
    | T7 | 951-1000 | Autonomous | Full autonomy, self-governance |

    ## Related Resources
    - [BASIS Specification](https://basis.vorion.org)
    - [TypeScript SDK](https://npmjs.com/package/@vorionsys/cognigate)
    - [Agent Anchor Trust](https://trust.agentanchorai.com)

  version: 1.0.0
  contact:
    name: Agent Anchor AI
    email: api-support@agentanchorai.com
    url: https://cognigate.dev
  license:
    name: MIT
...
  x-logo:
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /proofs/{proofId}:
    get:
      tags: [Proofs]
      summary: Get proof record
      description: Retrieve a specific proof record
      operationId: getProof
      parameters:
        - name: proofId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Proof record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProofRecordResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      required: [success, error, requestId, timestamp]
      properties:
        success:
          type: boolean
          const: false
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
        requestId:
          type: string
        timestamp:
          type: string
          format: date-time
...
    GovernanceDecision:
      type: string
      enum: [ALLOW, DENY, ESCALATE, DEGRADE]

    # ==========================================================================
    # AGENT SCHEMAS
    # ==========================================================================
...
    ProofRecord:
      type: object
      required: [id, entityId, intentId, decision, action, outcome, trustScoreBefore, trustScoreAfter, timestamp, hash, previousHash]
      properties:
        id:
          type: string
        entityId:
          type: string
        intentId:
          type: string
        decision:
          $ref: '#/components/schemas/GovernanceDecision'
        action:
          type: string
        outcome:
          type: string
          enum: [SUCCESS, FAILURE, PARTIAL, PENDING]
        trustScoreBefore:
          type: number
        trustScoreAfter:
          type: number
        timestamp:
          type: string
          format: date-time
        hash:
          type: string
        previousHash:
          type: string
        metadata:
          type: object
    ProofRecordResponse:
      type: object
      required: [success, data, requestId, timestamp]
      properties:
        success:
          type: boolean
          const: true
        data:
          $ref: '#/components/schemas/ProofRecord'
        requestId:
          type: string
        timestamp:
          type: string
          format: date-time
...